cgroup: move some functions to a new file
