+++ src/libcrun/cgroup-internal.h
@@ -37,4 +37,8 @@ int chown_cgroups (const char *path, uid_t uid, gid_t gid, libcrun_error_t *err)
 int destroy_cgroup_path (const char *path, int mode, libcrun_error_t *err);
 int cgroup_killall_path (const char *path, int signal, libcrun_error_t *err);
 
+int libcrun_cgroup_read_pids_from_path (const char *path, bool recurse, pid_t **pids, libcrun_error_t *err);
+
+bool read_proc_cgroup (char *content, char **saveptr, char **id, char **controller_list, char **path);
+
 #endif
