+++ tests/tests_libcrun_fuzzer.c
@@ -19,6 +19,8 @@
 #include <libcrun/error.h>
 #include <libcrun/cgroup.h>
 #include <libcrun/cgroup-systemd.h>
+#include <libcrun/cgroup-utils.h>
+#include <libcrun/cgroup-internal.h>
 #include <libcrun/utils.h>
 #include <libcrun/status.h>
 #include <libcrun/seccomp.h>
